requirements.txt
pyunpack==0.2.2
python-magic==0.4.25
tqdm==4.62.3